+++ lib/src/input_password.dart
@@ -48,6 +48,7 @@ class InputPassword extends InputField<String> {
     this.textInputAction,
     this.toolbarOptions,
     List<InputValidator> validators,
+    this.visibilityIcon = Icons.remove_red_eye,
     this.wantKeepAlive = false,
   })  : assert(autofocus != null),
         assert(autovalidate != null),
@@ -99,6 +100,8 @@ class InputPassword extends InputField<String> {
   final TextDirection textDirection;
   final TextInputAction textInputAction;
   final ToolbarOptions toolbarOptions;
+  /// Icon on button to make password visible. `null` disables the button.
+  final IconData visibilityIcon;
   @override
   final bool wantKeepAlive;
 
@@ -136,53 +139,56 @@ class _InputPasswordState<T> extends InputFieldState<String> {
     final InputDecoration effectiveDecoration =
         (widget.decoration ?? const InputDecoration())
             .applyDefaults(Theme.of(context).inputDecorationTheme);
+TextField textField =    TextField(
+    autocorrect: false,
+    autofocus: widget.autofocus,
+    buildCounter: widget.buildCounter,
+    controller: _effectiveController,
+    cursorColor: widget.cursorColor,
+    cursorRadius: widget.cursorRadius,
+    cursorWidth: widget.cursorWidth,
+    decoration: effectiveDecoration.copyWith(errorText: errorText),
+    enabled: super.isEnabled(),
+    enableInteractiveSelection: widget.enableInteractiveSelection,
+    enableSuggestions: false,
+    expands: widget.expands,
+    focusNode: widget.focusNode,
+    keyboardAppearance: widget.keyboardAppearance,
+    keyboardType:
+    _visible ? TextInputType.visiblePassword : TextInputType.text,
+    inputFormatters: widget.inputFormatters,
+    maxLength: widget.maxLength,
+    maxLengthEnforced: widget.maxLengthEnforced,
+    maxLines: 1,
+    minLines: 1,
+    obscureText: (!super.isEnabled()) || (!_visible),
+    onChanged: (v) => didChange(v),
+    onEditingComplete: widget.onEditingComplete,
+    onSubmitted: widget.onFieldSubmitted,
+    onTap: widget.onTap,
+    readOnly: false,
+    scrollPadding: widget.scrollPadding,
+    showCursor: widget.showCursor,
+    style: widget.style,
+    strutStyle: widget.strutStyle,
+    textAlign: widget.textAlign,
+    textAlignVertical: widget.textAlignVertical,
+    textCapitalization: widget.textCapitalization,
+    textDirection: widget.textDirection,
+    textInputAction: widget.textInputAction,
+    toolbarOptions: widget.toolbarOptions,
+    );
 
-    return Row(
+    return (widget.visibilityIcon == null ) ?
+    textField :
+      Row(
       children: [
         Flexible(
-          child: TextField(
-            autocorrect: false,
-            autofocus: widget.autofocus,
-            buildCounter: widget.buildCounter,
-            controller: _effectiveController,
-            cursorColor: widget.cursorColor,
-            cursorRadius: widget.cursorRadius,
-            cursorWidth: widget.cursorWidth,
-            decoration: effectiveDecoration.copyWith(errorText: errorText),
-            enabled: super.isEnabled(),
-            enableInteractiveSelection: widget.enableInteractiveSelection,
-            enableSuggestions: false,
-            expands: widget.expands,
-            focusNode: widget.focusNode,
-            keyboardAppearance: widget.keyboardAppearance,
-            keyboardType:
-                _visible ? TextInputType.visiblePassword : TextInputType.text,
-            inputFormatters: widget.inputFormatters,
-            maxLength: widget.maxLength,
-            maxLengthEnforced: widget.maxLengthEnforced,
-            maxLines: 1,
-            minLines: 1,
-            obscureText: (!super.isEnabled()) || (!_visible),
-            onChanged: (v) => didChange(v),
-            onEditingComplete: widget.onEditingComplete,
-            onSubmitted: widget.onFieldSubmitted,
-            onTap: widget.onTap,
-            readOnly: false,
-            scrollPadding: widget.scrollPadding,
-            showCursor: widget.showCursor,
-            style: widget.style,
-            strutStyle: widget.strutStyle,
-            textAlign: widget.textAlign,
-            textAlignVertical: widget.textAlignVertical,
-            textCapitalization: widget.textCapitalization,
-            textDirection: widget.textDirection,
-            textInputAction: widget.textInputAction,
-            toolbarOptions: widget.toolbarOptions,
-          ),
+          child: textField,
         ),
         IconButton(
           icon: Icon(
-            Icons.remove_red_eye,
+            widget.visibilityIcon,
             color:
                 (super.isEnabled() && _visible) ? Colors.green : Colors.black,
           ),